Removed isobath creation logic and moved to the HHM Simple Engine
